requirement.txt
cog==0.8.6
cpm_kernels==1.0.11
deepspeed==0.9.4
fairscale==0.4.13
huggingface_hub==0.14.1
intel_extension_for_pytorch==2.0.100
jieba==0.42.1
jsonlines==3.1.0
nevergrad==0.13.0
nltk==3.8.1
numpy==1.24.3
openai==0.28.0
optuna==3.3.0
packaging==23.2
pandas==2.0.1
ray==2.7.0
rouge==1.0.1
rouge_chinese==1.0.3
scikit_learn==1.3.0
sentencepiece==0.1.99
setuptools==67.8.0
tqdm==4.65.0
transformers==4.28.1
wandb==0.15.12
dataset
accelerate